Move logging to its own module.
